+++ backend/image_generator/templates/api.html
@@ -1,0 +1,1 @@
+<h1>temp</h1>
